FIX - map reduce treesPerPopulation

+++ api/src/main/java/itba/pod/api/mappers/CounterMapper.java
@@ -3,10 +3,10 @@ package itba.pod.api.mappers;
 import com.hazelcast.mapreduce.Context;
 import com.hazelcast.mapreduce.Mapper;
 
-public class CounterMapper<K, V> implements Mapper<K, V, V, Integer> {
+public class CounterMapper<K, V> implements Mapper<K, V, V, Long> {
 
     @Override
-    public void map(K key, V value, Context<V, Integer> context) {
-        context.emit(value, 1);
+    public void map(K key, V value, Context<V, Long> context) {
+        context.emit(value, 1L);
     }
 }
